sub folder removal and start.py merge to root
