--- ocv_projects-master/README.md
@@ -1,20 +1,0 @@
-# OpenCV Projects
-
-Run from command line with arguments to interface with the different functions that will be included in the system.
-
-## Done
-* Haar Casade face detection algorithm
-* Improved argument interface
-* Cut faces out of image
-* Webcam face detector
-
-
-## To Do
-* Eye detection once face has been found
-* Save faces out of image, whereto argument? 
-* Red eye removal?
-* Other facial features detected
-* Train NN to identify faces
-* Better webcam speed
-* Pedestrian counter and face extraction 
-
