+++ start.py
@@ -41,6 +41,30 @@ def faces(imageScan, cascadePath, toCut):
     
   cv2.imshow("Faces found",image)
   cv2.waitKey(0)
+
+def videoCapture(xml, camera):
+
+	video = cv2.VideoCapture(camera)
+	cascade = cv2.CascadeClassifier(xml)
+
+	while(video.isOpened()):
+		returnValue, frame = video.read()
+
+		cv2.imshow("Original", frame)
+		
+		facesFound = cascade.detectMultiScale(frame,1.3,5)  
+
+  		print("Found {} faces!".format(len(facesFound)))
+
+  		for (x,y,w,h) in facesFound:
+    			cv2.rectangle(image, (x,y),(x+w, y+h), (0,255,0), 2)
+
+		cv2.imshow("Face Find", frame)
+
+		if key == 27: # exit on ESC
+        		break
+	video.release()
+	cv2.destroyAllWindows()
 #------------------------------------------------------------------------------
 # Main
 #------------------------------------------------------------------------------
@@ -50,6 +74,7 @@ parser.add_argument("image", help="file path to image for processing")
 parser.add_argument("xml_path", help="path to XML")
 parser.add_argument("-c","--cut", help="cut faces from image",
  action="store_true")
+parse.add_argument("-C","--Camera", type=int, help="Camera that will capture video, default 0")
 
 args = parser.parse_args()
 
@@ -61,3 +86,8 @@ except IOError:
 
 
 faces(args.image, args.xml_path, args.cut)
+
+if args.camera != 0:
+	videoCapture(args.xml_path, args.camera)
+else:	
+	videoCapture(args.xml_path, 0)
